Kernel - Cleaning up
- Decided against PAE support for the 32-bit x86 code
- Some debug in System_*
- Still designing parts of mmap

+++ Kernel/arch/x86/include/arch.h
@@ -87,7 +87,7 @@ typedef signed long long	Sint64;
 typedef Uint	size_t;
 typedef char	BOOL;
 
-typedef Uint64	tPAddr;
+typedef Uint32	tPAddr;
 typedef Uint32	tVAddr;
 
 typedef struct {
